% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioinfoRtools-package.R
\docType{package}
\name{BioinfoRtools-package}
\alias{BioinfoRtools}
\alias{BioinfoRtools-package}
\title{BioinfoRtools: A comprehensive bioinformatics R package toolbox}
\description{
A comprehensive bioinformatics R package toolbox.
}
\author{
\strong{Maintainer}: Yulong Qin \email{524583919@qq.com} [copyright holder]

Other contributors:
\itemize{
  \item Yulong Qin \email{524583919@qq.com} [contributor]
}

}
\keyword{internal}
