% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinic_corr.R
\name{Clinic_corr}
\alias{Clinic_corr}
\title{Calculate correlations between continuous and discrete variables}
\usage{
Clinic_corr(
  continuous = NULL,
  continuous2 = NULL,
  binary = NULL,
  grp_nm = "Clinic_corr1",
  dir_nm = "Clinic_corr",
  skip_cont = F,
  skip_bina = F,
  skip_cont_bina = F
)
}
\arguments{
\item{continuous}{the continuous data.frame}

\item{continuous2}{the continuous2 data.frame}

\item{binary}{the binary data.frame}

\item{grp_nm}{the name of the group}

\item{dir_nm}{the name of the directory}

\item{skip_cont}{whether to skip continuous to continuous}

\item{skip_bina}{whether to skip binary to binary}

\item{skip_cont_bina}{whether to skip continuous to binary}
}
\description{
Calculate correlations between continuous and discrete variables
}
\details{
Calculate correlations between continuous and discrete variables
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/Clinic_corr.R")
Clinic_corr(
  continuous = NULL, binary = NULL,
  grp_nm = "Clinic_corr1", dir_nm = "Clinic_corr",
  skip_cont = F, skip_bina = F
)
}
}
\seealso{
\code{\link{GO_KEGG}}, \code{\link{GSEA_GO_KEGG}}, \code{\link{GSVA_ssGSEA}}
}
\author{
Yulong Qin
}
