% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coldata}
\alias{coldata}
\title{Meta-data of samples}
\format{
A data frame with 7 rows and 2 variables:
\describe{
\item{condition}{condition}
\item{type}{type}
}
}
\usage{
coldata
}
\description{
Meta-data of a sample of 7 samples and 2 clinical information samples.
}
\keyword{datasets}
