% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countdata}
\alias{countdata}
\title{Fuel economy data from 1999 to 2008 for 38 popular models of cars}
\format{
A data frame with 234 rows and 11 variables:
\describe{
\item{treated1}{manufacturer name}
\item{treated2}{model name}
\item{treated3}{engine displacement, in litres}
\item{untreated1}{year of manufacture}
\item{untreated2}{number of cylinders}
\item{untreated3}{type of transmission}
\item{untreated4}{type of transmission}
}
}
\usage{
countdata
}
\description{
This dataset contains a subset of the fuel economy data that the EPA makes
available on \url{https://fueleconomy.gov/}. It contains only models which
had a new release every year between 1999 and 2008 - this was used as a
proxy for the popularity of the car.
}
\keyword{datasets}
