% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGs_Wilcoxon.R
\name{DEGs_Wilcoxon}
\alias{DEGs_Wilcoxon}
\title{Using Wilcoxon for DEGs analysis}
\usage{
DEGs_Wilcoxon(
  countdata,
  group,
  case = "case",
  ctrl = "ctrl",
  covariate = NULL,
  updown_thres = 1,
  rate_sap = 0.3,
  OUT_df_index = FALSE,
  updown_index = FALSE,
  p_value = 0.05,
  padj_value = 0.05,
  grp_nm = "DEGs_Wilcoxon",
  dir_nm = "DEGs_Wilcoxon"
)
}
\arguments{
\item{countdata}{expression matrix}

\item{group}{the group of samples}

\item{case}{the case of samples}

\item{ctrl}{the control of samples}

\item{covariate}{the covariate of samples}

\item{updown_thres}{the threshold of updown}

\item{rate_sap}{the rate of sap}

\item{OUT_df_index}{the index of output df}

\item{updown_index}{the index of updown}

\item{p_value}{the p value}

\item{padj_value}{the padj value}

\item{grp_nm}{the name of group}

\item{dir_nm}{the name of dir}
}
\value{
data.frame
}
\description{
This function uses Wilcoxon to identify differentially expressed genes (DEGs)
between two groups of samples. It takes in gene count data and returns a list of
DEGs with adjusted p-values and log2 fold changes.
}
\details{
Using Wilcoxon for DEGs analysis
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/DEGs_Wilcoxon.R")
df <- DEGs_Wilcoxon(countdata, group,
  case = "case", ctrl = "ctrl", rate_sap = 0.1,
  OUT_df_index = T, updown_index = T
)
}
}
\seealso{
\code{\link{DEGs_DESeq2}}
}
\author{
Yulong Qin
}
