% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_KEGG.R
\name{GO_KEGG}
\alias{GO_KEGG}
\title{GO and KEGG analysis}
\usage{
GO_KEGG(
  updown = NULL,
  up = NULL,
  down = NULL,
  simplify_index = F,
  maxGSSize = 500,
  grp_nm = "GO_KEGG",
  dir_nm = "GO_KEGG"
)
}
\arguments{
\item{updown}{updown genes}

\item{up}{up genes}

\item{down}{down genes}

\item{simplify_index}{whether simplify}

\item{maxGSSize}{max gene set size}

\item{grp_nm}{the name of group}

\item{dir_nm}{the name of dir}
}
\description{
GO and KEGG analysis
}
\details{
GO and KEGG analysis
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/GO_KEGG.R")
GO_KEGG(
  updown = gene_bf_go_updown, up = gene_bf_go_up, down = gene_bf_go_down,
  simplify_index = F, grp_nm = "GO_KEGG", dir_nm = "GO_KEGG"
)
}

}
\seealso{
\code{\link{GO_KEGG}},\code{\link{GSEA_GO_KEGG}}, \code{\link{GSVA_ssGSEA}}
}
\author{
Yulong Qin
}
