% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_GO_KEGG.R
\name{GSEA_GO_KEGG}
\alias{GSEA_GO_KEGG}
\title{GSEA analysis}
\usage{
GSEA_GO_KEGG(
  DEGs = NULL,
  simplify_index = F,
  maxGSSize = 500,
  grp_nm = "GSEA_GO_KEGG",
  dir_nm = "GSEA_GO_KEGG"
)
}
\arguments{
\item{DEGs}{the input DEGs must be a df with only two columns of c('SYMBOL','LogFC')}

\item{simplify_index}{whether to simplify the results}

\item{maxGSSize}{the maximum number of genes in the gene set}

\item{grp_nm}{the name of the group}

\item{dir_nm}{the name of the directory}
}
\description{
GSEA analysis
}
\details{
GSEA analysis
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/GSEA_GO_KEGG.R")
GSEA_GO_KEGG(
  DEGs = NULL, simplify_index = F,
  grp_nm = "GSEA_GO_KEGG", dir_nm = "GSEA_GO_KEGG"
)
}
}
\seealso{
\code{\link{GO_KEGG}},\code{\link{GSEA_GO_KEGG}}, \code{\link{GSVA_ssGSEA}}
}
\author{
Yulong Qin
}
