% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA_ssGSEA.R
\name{GSVA_ssGSEA}
\alias{GSVA_ssGSEA}
\title{GSVA and ssGSEA analysis}
\usage{
GSVA_ssGSEA(
  exp_df = NULL,
  gsva_mat_go_scale = NULL,
  gsva_mat_kegg_scale = NULL,
  ssgsea_mat_go_scale = NULL,
  ssgsea_mat_kegg_scale = NULL,
  group = NULL,
  case = "Case",
  ctrl = "Ctrl",
  path_gmt_GO = "./inputdata/c5.go.v2023.1.Hs.symbols_BP.gmt",
  path_gmt_KEGG = "./inputdata/c5.go.v2023.1.Hs.symbols_BP.gmt",
  grp_nm = "GSVA_ssGSEA",
  dir_nm = "GSVA_ssGSEA",
  OUT_df_index = FALSE,
  updown_index = FALSE,
  updown_thres = 1,
  p_value = 0.05,
  padj_value = 0.05
)
}
\arguments{
\item{exp_df}{TPM matrix}

\item{gsva_mat_go_scale}{the result of GSVA analysis, gsva_mat_go_scale}

\item{gsva_mat_kegg_scale}{the result of GSVA analysis, gsva_mat_kegg_scale}

\item{ssgsea_mat_go_scale}{the result of ssGSEA analysis, ssgsea_mat_go_scale}

\item{ssgsea_mat_kegg_scale}{the result of ssGSEA analysis, ssgsea_mat_kegg_scale}

\item{group}{the group of samples}

\item{case}{the name of case}

\item{ctrl}{the name of ctrl}

\item{path_gmt_GO}{the path of GO gmt file}

\item{path_gmt_KEGG}{the path of KEGG gmt file}

\item{grp_nm}{the name of group}

\item{dir_nm}{the name of directory}

\item{OUT_df_index}{whether to output the index of the result}

\item{updown_index}{whether to output the index of up and down}

\item{updown_thres}{the threshold of up and down}

\item{p_value}{the threshold of p value}

\item{padj_value}{the threshold of padj value}
}
\description{
GSVA and ssGSEA analysis
}
\details{
GSVA and ssGSEA analysis
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/GSVA_ssGSEA.R")
GSVA_ssGSEA(
  exp_df = exp_df,
  group = group,
  case = "Case", ctrl = "Ctrl",
  grp_nm = "GSVA_ssGSEA", dir_nm = "GSVA_ssGSEA",
  OUT_df_index = FALSE, updown_index = FALSE, updown_thres = 1,
  p_value = 0.05, padj_value = 0.05
)
}
}
\seealso{
\code{\link{GO_KEGG}},\code{\link{GSEA_GO_KEGG}}, \code{\link{GSVA_ssGSEA}}
}
\author{
Yulong Qin
}
