% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_set.R
\name{manipulate-set}
\alias{manipulate-set}
\alias{get_intersection}
\alias{get_union}
\alias{get_intersection_list}
\alias{get_union_list}
\title{computes the intersection, union, and difference sets between sets}
\usage{
get_intersection(...)

get_union(...)

get_intersection_list(...)

get_union_list(...)
}
\arguments{
\item{...}{the input vectors}
}
\value{
a list of vectors
}
\description{
computes the intersection, union, and difference sets between sets
}
\details{
computes the intersection, union, and difference sets between sets
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/manipulate_set.R")
get_intersection(a, b, c, d)
}

}
\seealso{
\code{\link{shorten_names}},\code{\link{split-paste}},
\code{\link{pubmed_search}},\code{\link{manipulate-set}}
}
\author{
Yulong Qin
}
\concept{manipulate-set}
