% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfuzz_pipeline.R
\name{Mfuzz_pipeline}
\alias{Mfuzz_pipeline}
\title{Mfuzz analysis}
\usage{
Mfuzz_pipeline(
  exp = NULL,
  grp = NULL,
  cluster_num = 20,
  min_acore = 0.3,
  GO_KEGG = F,
  grp_nm = "Mfuzz",
  dir_nm = "Mfuzz"
)
}
\arguments{
\item{exp}{the input expression matrix}

\item{grp}{the input group}

\item{cluster_num}{the number of cluster}

\item{min_acore}{the min acore of cluster}

\item{GO_KEGG}{whether to do GO and KEGG analysis}

\item{grp_nm}{the name of group}

\item{dir_nm}{the name of directory}
}
\description{
Mfuzz analysis
}
\details{
Mfuzz analysis
}
\examples{
\dontrun{
  source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/Mfuzz_pipeline.R")
  Mfuzz_pipeline(exp = exp,grp = grp,
              cluster_num = 20,min_acore = 0.3,
              GO_KEGG = F,
              grp_nm = "Mfuzz",dir_nm = "Mfuzz")
              }
}
\seealso{
\code{\link{WGCNA_pipeline}}, \code{\link{miRNA_mRNA}}, \code{\link{Mfuzz_pipeline}}
}
\author{
Yulong Qin
}
