% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_mRNA.R
\name{miRNA_mRNA}
\alias{miRNA_mRNA}
\title{miRNA and mRNA predict}
\usage{
miRNA_mRNA(
  mRNA_genes = NULL,
  miRNA_genes = NULL,
  table_nm = "validated",
  type_nm = "miup_mdown",
  search_index = F,
  local_merge_index = F,
  search_nm = "[TIAB] AND (sperm[TIAB] OR semen[TIAB])",
  grp_nm = "group",
  dir_nm = "miRNA"
)
}
\arguments{
\item{mRNA_genes}{the input mRNA genes}

\item{miRNA_genes}{the input miRNA genes}

\item{table_nm}{the table name}

\item{type_nm}{the type name}

\item{search_index}{whether to search pubmed}

\item{local_merge_index}{whether to merge local data}

\item{search_nm}{the search name}

\item{grp_nm}{the group name}

\item{dir_nm}{the dir name}
}
\value{
the predict result
}
\description{
miRNA and mRNA predict
}
\details{
miRNA and mRNA predict
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/miRNA_mRNA.R")
miRNA_mRNA(mRNA_genes = NULL, # 输入up/down,软件不能区分mRNA-miRNA间的对应关系，只能通过输入的updown控制
          miRNA_genes = NULL, # 输入up/down
          table_nm = "validated", type_nm="miup_mdown",
          search_index = F,local_merge_index=F,
          search_nm = "[TIAB] AND (sperm[TIAB] OR semen[TIAB])",
          grp_nm = "group",dir_nm = "miRNA")
          }
}
\seealso{
\code{\link{WGCNA_pipeline}}, \code{\link{miRNA_mRNA}}, \code{\link{Mfuzz_pipeline}}
}
\author{
Yulong Qin
}
