% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA plots are drawn according to groups}
\usage{
PCA(countdata, group, grp_nm = "PCA", scale = T)
}
\arguments{
\item{countdata}{the count data matrix}

\item{group}{the group information}

\item{grp_nm}{the name of the group}

\item{scale}{whether to scale the data}
}
\description{
PCA plots are drawn according to groups
}
\details{
PCA plots are drawn according to groups
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/PCA.R")
df <- iris[1:4] \%>\%
  t() \%>\%
  as.data.frame()
PCA(df, iris$Species, scale = F)
}
}
\seealso{
\code{\link{PCA}},\code{\link{HTmap}}
}
\author{
Yulong Qin
}
