% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{pubmed_search}
\alias{pubmed_search}
\title{serach pubmed by keywords}
\usage{
pubmed_search(
  search_words = NULL,
  fixed_words = NULL,
  sep = "AND",
  num_article = 20,
  api_key = "54d30eb11bfb539cfbab47558615ccf9e809",
  grp_nm = "pubmed_search",
  dir_nm = "pubmed_search"
)
}
\arguments{
\item{search_words}{keywords}

\item{fixed_words}{fixed words}

\item{sep}{the sep of keywords and fixed words}

\item{num_article}{the number of article}

\item{api_key}{the api_key of pubmed}

\item{grp_nm}{the group name}

\item{dir_nm}{the dir name}
}
\value{
my_abstracts_txt
}
\description{
serach pubmed by keywords
}
\details{
serach pubmed by keywords
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/pubmed_search.R")
pubmed_search(
  search_words = NULL, fixed_words = NULL, sep = "AND",
  api_key = "54d30eb11bfb539cfbab47558615ccf9e809",
  grp_nm = "pubmed_search", dir_nm = "pubmed_search"
)
}

}
\seealso{
\code{\link{shorten_names}}, \code{\link{split_paste}},
\code{\link{pubmed_search}}, \code{\link{manipulate_set}}
}
\author{
Yulong Qin
}
