% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_names.R
\name{shorten_names}
\alias{shorten_names}
\title{Shorten the description of excessively long GO enrichment}
\usage{
shorten_names(x, n_word = 6, n_char = 40)
}
\arguments{
\item{x}{character vector}

\item{n_word}{number of words}

\item{n_char}{number of characters}
}
\value{
character vector
}
\description{
Shorten the description of excessively long GO enrichment
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/shorten_names.R")
go_res$Des_short <- sapply(go_res$Description, shorten_names)
}

}
\seealso{
\code{\link{shorten_names}}, \code{\link{split_paste}},
\code{\link{pubmed_search}}, \code{\link{manipulate_set}}
}
\author{
Yulong Qin
}
