% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_paste.R
\name{split-paste}
\alias{split-paste}
\alias{split_paste}
\title{convert the string separator}
\usage{
split_paste(string, split = " ", collapse = "\\t")
}
\arguments{
\item{string}{the string}

\item{split}{the split}

\item{collapse}{the collapse}
}
\value{
a string
}
\description{
convert the string separator
}
\details{
convert the string separator
}
\examples{
\dontrun{
a <- "a_b_c_d_e"
split_paste(a, split = "_", collapse = "-")
}

}
\seealso{
\code{\link{shorten_names}},\code{\link{split-paste}},
\code{\link{pubmed_search}},\code{\link{manipulate-set}}
}
\author{
Yulong Qin
}
