% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_pipeline.R
\name{WGCNA_pipeline}
\alias{WGCNA_pipeline}
\title{the pipeline of WGCNA}
\usage{
WGCNA_pipeline(
  exp = NULL,
  clinic = NULL,
  num_disc = NA,
  strict = T,
  threads = 8,
  rate_sap = 0.3,
  num_genes = 5000,
  minModule = 30,
  CutHeight = 0.25,
  deepSplit = 2,
  grp_nm = "WGCNA_try1",
  dir_nm = "WGCNA"
)
}
\arguments{
\item{exp}{the input expression matrix}

\item{clinic}{the input clinic}

\item{num_disc}{the number of discrete variables}

\item{strict}{whether to strict}

\item{threads}{the number of threads}

\item{rate_sap}{the rate of sample}

\item{num_genes}{the number of genes}

\item{minModule}{the number of minModule}

\item{CutHeight}{the height of cut}

\item{deepSplit}{the deep of split}

\item{grp_nm}{the group name}

\item{dir_nm}{the dir name}
}
\description{
the pipeline of WGCNA
}
\details{
the pipeline of WGCNA
}
\examples{
\dontrun{
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/WGCNA_pipeline.R")
WGCNA_pipeline(exp = NULL, # 输入原始的TPM即可，行为基因，列为样本
              clinic = NULL,  # 行为样本，列为变量，离散变量在前，连续变量在后
              num_disc = NA, strict = T,threads = 8, # 一般默认值就可以
              rate_sap = 0.3, num_genes = 5000, # 一般默认值就可以
              minModule=30,CutHeight=0.25,deepSplit = 2,
              grp_nm = "WGCNA_try1",dir_nm = "WGCNA")
              }

}
\seealso{
\code{\link{WGCNA_pipeline}}, \code{\link{miRNA_mRNA}}, \code{\link{Mfuzz_pipeline}}
}
\author{
Yulong Qin
}
