% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coldata}
\alias{coldata}
\title{Prices of over 50,000 round cut diamonds}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{condition}{price in US dollars ($326--$18,823)}
\item{type}{weight of the diamond (0.2--5.01)}
}
}
\usage{
coldata
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\keyword{datasets}
