% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countdata}
\alias{countdata}
\title{A transcriptome expression}
\format{
A data frame with 14599 rows and 7 variables:
\describe{
\item{treated1}{treated1}
\item{treated2}{treated2}
\item{treated3}{treated3}
\item{untreated1}{untreated1}
\item{untreated2}{untreated2}
\item{untreated3}{untreated3}
\item{untreated4}{untreated4}
}
}
\usage{
countdata
}
\description{
A transcriptome expression profile of 14599 genes in 7 samples.
}
\keyword{datasets}
