% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\alias{PCA.data.frame}
\title{PCA plots are drawn according to groups}
\usage{
PCA(countdata, group, scale_index = T, method = "PCA", grp_nm = "PCA")

\method{PCA}{data.frame}(countdata, group, scale_index = T, method = "PCA", grp_nm = "PCA")
}
\arguments{
\item{countdata}{the count data matrix}

\item{group}{the group information}

\item{scale_index}{whether to scale the data}

\item{method}{the method of dimensionality reduction}

\item{grp_nm}{the name of the group}
}
\value{
plot
}
\description{
PCA plots are drawn according to groups
}
\details{
PCA plots are drawn according to groups

this function use \code{ggplot()} to draw PCA plots according to groups
}
\section{Useful mutate functions}{

\itemize{
\item \code{\link{+}}, \code{\link{-}}, \code{\link[=log]{log()}}, etc., for their usual mathematical meanings
}

...
}

\examples{
\dontrun{

# this is a example
source(file = "F:/Bioinformatic_repository/02_R/code_R/A_Script_Function/PCA.R")
df <- iris[1:4] \%>\% t() \%>\% as.data.frame()
PCA(df,iris$Species,scale_index = T
PCA(df,iris$Species,method = "UMAP",scale_index = T)
PCA(df,iris$Species,method = "tSNE",scale_index = T)
PCA(df,iris$Species,method = "LDA",scale_index = T)
PCA(df,iris$Species,method = "lda",scale_index = T)
}

}
\seealso{
\code{\link{PCA}}, \code{\link{HTmap}}

\code{\link[=ggplot]{ggplot()}}, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} which this function wraps.
}
\author{
Yulong Qin
}
